pyproject.toml
[project]
name = "zuri-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.7",
    "aiohttp>=3.12.15",
    "langgraph>=0.6.4",
    "colorama>=0.4.6",
    "pandas>=2.3.1",
    "pip>=25.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "google-cloud-storage==2.19.0",
    "pymupdf>=1.26.3",
    "pillow>=11.3.0",
]
